refactor: rename test-functions to testFunctions
